Update Hugo baseURL
This CL updates the Hugo `baseURL` to match the
`google.github.io/tour-of-wgsl` URL.

The title in Hugo and in the header are also updated to `Tour of WGSL`.
The justification on the header menu is updated to `center` when the
screen is larger which makes any word wrapping which can happen look a
bit nicer.