Update pages to use {{< pages >}} short code
A few of the index pages were still rendering sections as a `ul`. The
URLs were not updated correctly when moving to the final hosting
location so the links did not work.

This CL updates those index pages to use the short code generator
instead of a hard coded list.

Fixes #31